pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ma-pipeline"              
version = "0.1.0"                  
description = "Offline RAG + multi-agent pipeline."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2,<3",
  "python-dotenv>=1,<2",
  "langchain>=1,<2",
  "langchain-community>=0.4,<1",
  "langchain-huggingface>=1,<2",
  "langchain-text-splitters>=0.3,<1",
  "sentence-transformers>=2,<3",
  "faiss-cpu>=1,<2",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[tool.setuptools]
# Treat as packages ( __init__.py)
packages = ["src", "data", "rag", "agents", "ui"]